fix promisified fs.exists
fs.exists callback is not being called with an error as most node callbacks are.
This prevents bluebird.promisify from working as expected.
The boolean passed into the callback that indicates if the file exists
is being misinterpreted as an error, causing the promise always be rejected.

```js
// pseudo implementation of exists:
function exists(file, callback) {
  // ... lets assume file exists
  callback(true);
}

// kind of what promisify does internaly:
exists('foo.js', function(err, args...) {
  if (err != null) {
    reject(err);
  } else {
    resolve(args...);
  }
});
```

Documentation of fs.exists:
https://nodejs.org/api/fs.html#fs_fs_exists_path_callback

fix #5